Working CRUD

+++ backend/superbikes/src/main/java/com/tiburcio/superbikes/entity/dao/IBicycleDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.tiburcio.superbikes.entity.models.Bicycle;
 
-public interface IBicycleDao extends CrudRepository<Bicycle, Integer> {
+public interface IBicycleDao extends CrudRepository<Bicycle, Long> {
 	
 }
